refactor(analysis)
